Fix running time performance
An unwanted side-effect of fixing the space leak was the the importer wasn't leverage concurrency as desired: when parsing a batch of entries it was waiting for the merge request to the database to be completed before executing the next batch. This cause a degradation of performance that was quite dramatic. In order to fix this, now the importer uses a job queue internally where batches of entries ready to be executed are left so that the next available worker can execute them.